Extract helper functions to helpers.go module
